Use zero-config conventions
