+++ api/_examples/http-server-api.ts
@@ -1,5 +1,5 @@
 /**
- * API Demo of adding WebSocketOverHttp support to any http.RequestListener function (e.g. (req, res) => void).
+ * API Demo of adding WebSocketOverHttp support patching any http.Server
  * Almost all node.js web libraries support creating one of these from the underlying Application object.
  * In this example, we use zeit/micro, but you can do something similar with koa, express, raw node http, etc.
  */
@@ -9,15 +9,13 @@ import {
   IStoredConnection,
   IStoredPubSubSubscription,
 } from "fanout-graphql-tools";
-import {
-  GraphqlWsOverWebSocketOverHttpRequestListener,
-  MapSimpleTable,
-} from "fanout-graphql-tools";
+import { MapSimpleTable } from "fanout-graphql-tools";
+import { GraphqlWsOverWebSocketOverHttpSubscriptionHandlerInstaller } from "fanout-graphql-tools";
 import * as http from "http";
-import { run as microRun } from "micro";
+import micro from "micro";
 import FanoutGraphqlApolloConfig, {
   FanoutGraphqlGripChannelsForSubscription,
-} from "../FanoutGraphqlApolloConfig";
+} from "../_lib/FanoutGraphqlApolloConfig";
 
 /**
  * WebSocket-Over-HTTP Support requires storage to keep track of ws-over-http connections and subscriptions.
@@ -40,27 +38,20 @@ const apolloServerConfig = FanoutGraphqlApolloConfig({
 });
 const apolloServer = new ApolloServer(apolloServerConfig);
 
+// Note: In micro 9.3.5 this will return an http.RequestListener instead (after https://github.com/zeit/micro/pull/399)
+// Provide it to http.createServer to create an http.Server
+const httpServer: http.Server = micro(apolloServer.createHandler());
+
 // This won't throw, but it also won't result in working WebSocket Subscriptions (when you create the subscription via gql api, a response comes back mentioning:
 // { "error": { "name": "TypeError", "message": "Cannot read property 'addListener' of undefined" }
 // But there is nothing useful on stderr of the server.
 // apolloServer.installSubscriptionHandlers(httpServer)
-
-// In micro 9.3.5, the default export of micro(handler) will return an http.RequestListener (after https://github.com/zeit/micro/pull/399).
-// As of this authoring, only 9.3.4 is out, which returns an http.Server. So we manually build the RequestListner here.
-// After 9.3.5, the following will work:
-// import micro from "micro"
-// const microRequestListener = micro(apolloServer.createHandler())
-const microRequestListener: http.RequestListener = (req, res) =>
-  microRun(req, res, apolloServer.createHandler());
-
-const httpServer = http.createServer(
-  GraphqlWsOverWebSocketOverHttpRequestListener(microRequestListener, {
-    connectionStorage: webSocketOverHttpStorage.connections,
-    getGripChannel: FanoutGraphqlGripChannelsForSubscription,
-    pubSubSubscriptionStorage: webSocketOverHttpStorage.pubSubSubscriptions,
-    schema: apolloServerConfig.schema,
-  }),
-);
+GraphqlWsOverWebSocketOverHttpSubscriptionHandlerInstaller({
+  connectionStorage: webSocketOverHttpStorage.connections,
+  getGripChannel: FanoutGraphqlGripChannelsForSubscription,
+  pubSubSubscriptionStorage: webSocketOverHttpStorage.pubSubSubscriptions,
+  schema: apolloServerConfig.schema,
+})(httpServer);
 
 const port = process.env.PORT || 57410;
 httpServer.listen(port, () => {
