+++ api/index.ts
@@ -1,0 +1,5 @@
+import { main } from './_lib/FanoutGraphqlExpressServer';
+
+main().catch(error => {
+	throw error;
+});
