+++ now.json
@@ -1,9 +1,9 @@
 {
   "version": 2,
   "builds": [
-    { "src": "api/FanoutGraphqlExpressServer.ts", "use": "@now/node"}
+    { "src": "api/index.ts", "use": "@now/node"}
   ],
   "routes": [
-    { "src": ".*", "dest": "/api/FanoutGraphqlExpressServer.ts" }
+    { "src": ".*", "dest": "/api/index.ts" }
   ]
 }
